Module Ppx_deriving_binary_bytes__Variant

type constuctor_arguments = [

| `TUPLE of Ppxlib.core_type list

| `RECORD of Ppxlib.label_declaration list

]

type constructor =

| TAGGED of {

name : string;

args : constuctor_arguments;

value : Ppxlib.constant;

loc : Ppxlib.Location.t;

}

| ELSE of {

name : string;

args : constuctor_arguments;

loc : Ppxlib.Location.t;

}

val assert_not_GADT : Ppxlib.type_declaration -> Ppxlib.constructor_declaration list -> unit

val get_value : deriver:string -> int -> Ppxlib.attributes -> Ppxlib.constant

val get_else_flag : deriver:string -> Ppxlib.attributes -> bool

val constructors_of_ocaml_constructors : deriver:string -> Ppxlib.constructor_declaration list -> constructor list

Converts a list of OCaml constructors into a list of constructors in this module.

val constructors_of_ocaml_row_fields : deriver:string -> Ppxlib.row_field list -> constructor list

Converts a list of OCaml row_fields into a list of constructors in this module.

val invariant_constructors : constructor list -> unit