Module Ppx_deriving_binary_bytes_runtime__PrefixedList

Prefixed lists

('a, 'b) PrefixedList.t is a type of lists of elements of a type 'a, and a prefix of a type 'b.

For example, [0x00112233l; 0x44556677l; 0x8899aabb] of a type (uint32be, uint16be) PrefixedList.t is represented as

the heading two bytes for the length of a list (0x0003), and

the 4x3 bytes for contents of a list.

00  03  00  11  22  33  44  55  66  77  88  99  aa  bb
<---->  <--- x[0] --->  <--- x[1] --->  <--- x[2] --->
# of elements

type ('a, 'b) t = 'a list

val of_binary_bytes : ('a -> 'b -> 'c * 'b) -> ('a -> 'd -> int * 'b) -> 'a -> 'd -> 'c list * 'b

val binary_bytes_of : ('a -> 'b -> unit) -> ('a -> int -> unit) -> 'a -> 'b list -> unit